# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Protocol Agnostic Zero Trust Identity Broker",
        "title": "Shyntr Identity Hub API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/audit/{tenant_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves paginated audit logs for a specific tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "Get Tenant Audit Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID for isolation boundary",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of records to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tenant audit logs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "error - Failed to get audit logs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/t/{tenant_id}/oidc/callback": {
            "get": {
                "description": "Handles the callback from an external OIDC Identity Provider. Validates state and CSRF, exchanges the authorization code for tokens, fetches UserInfo, maps attributes, and resumes the original login flow.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "OIDC Federation"
                ],
                "summary": "OIDC Federation Callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code returned by the external IdP",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State parameter to mitigate CSRF and restore context",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects back to the original protocol handler (SAML or OAuth2) with a login_verifier",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid or missing callback parameters)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (CSRF validation failed, invalid state, or session expired)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error (connection not found or mapping failed)",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    }
                }
            }
        },
        "/t/{tenant_id}/oidc/login/{connection_id}": {
            "get": {
                "description": "Starts the login flow with an external OIDC Identity Provider (e.g., Google, Okta) for a specific tenant and connection. Generates CSRF tokens and redirects the user.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "OIDC Federation"
                ],
                "summary": "Initiate OIDC Federation Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OIDC Connection ID",
                        "name": "connection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The active cryptographic login challenge ID",
                        "name": "login_challenge",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects the user agent to the external Identity Provider's authorization endpoint",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request (missing tenant_id or login_challenge)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found (login request not found)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error (failed to initiate OIDC)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "payload.AppError": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "user_message": {
                    "type": "string",
                    "example": "Invalid input provided"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}